tests/monorepo-01/apps/payment-worker/requirements.txt
# VULNERABILITY: CVE-2020-1747 (Severity: CRITICAL)
# Arbitrary code execution via unsafe load
PyYAML==5.3
